refine structure and add macros
