--- klefki/src/types/algebra/concrete.rs
@@ -1,4 +1,0 @@
-//use crate::constrant;
-//use crate::types::algebra::{Field, GeneralField, GeneralGroup};
-//use std::cmp::{Ord, PartialEq, PartialOrd};
-//use std::ops::{Add, Div, Mul, Neg, Sub};
