+++ klefki/src/types/algebra/macros.rs
@@ -32,5 +32,72 @@ macro_rules! int_to_complex {
             }
         }
     )*}
+}
 
+macro_rules! arith_binary_self {
+    (
+        $Big:ty, $BigName:ident;
+        $($Imp:ident { $method:ident, $func:expr };
+        $ImpAssign:ident {$method_assign:ident};)*
+    ) => {
+        $(impl $Imp<$Big> for $Big {
+            type Output = $Big;
+            #[inline]
+            fn $method(self, other: $Big) -> $Big {
+                let incomplete = $func(self, other);
+                $BigName::from(incomplete)
+            }
+        }
+
+        impl $ImpAssign<$Big> for $Big {
+            #[inline]
+            fn $method_assign(&mut self, rhs: $Big) {
+                *self = {
+                    let cloned = self.clone();
+                    cloned.$method(rhs)
+                };
+            }
+        }
+
+        impl $Imp<&$Big> for $Big {
+            type Output = $Big;
+            #[inline]
+            fn $method(self, other: &$Big) -> $Big {
+                let other = other.clone();
+                self.$method(other)
+            }
+        }
+
+        impl $ImpAssign<&$Big> for $Big {
+            #[inline]
+            fn $method_assign(&mut self, other: &$Big) {
+                let other = other.clone();
+                self.$method_assign(other);
+            }
+        }
+        )*
+    };
+}
+
+macro_rules! arith_unary {
+    (
+        $Big:ty, $BigName:ident;
+        $Imp:ident {$method: ident, $func: expr};
+    ) => {
+        impl $Imp for $Big {
+            type Output = $Big;
+            fn $method(self) -> $Big {
+                let incomplete = $func(self);
+                $BigName::from(incomplete)
+            }
+        }
+
+        impl<'a> $Imp for &'a $Big {
+            type Output = $Big;
+            fn $method(self) -> $Big {
+                let other = self.clone();
+                other.$method()
+            }
+        }
+    };
 }
